# TicketingSystem/ticketing-system/src/app/api/models/create-news.ts
/* tslint:disable */
/* eslint-disable */
export interface CreateNews {
  content?: null | string;
  detailsTitle?: null | string;
  id?: number;
  image?: null | string;
  mainTitle?: null | string;
  publishDate?: Date;
  subTitle?: null | string;
  urlId?: null | string;
}