# TicketingSystem/ticketing-system/src/app/api/models/game-header.ts
/* tslint:disable */
/* eslint-disable */
export interface GameHeader {
  awayTeam?: null | string;
  awayTeamLogoUrl?: null | string;
  homeTeam?: null | string;
  homeTeamLogoUrl?: null | string;
  id?: number;
  place?: null | string;
  startTime?: Date;
}