# TicketingSystem/ticketing-system/src/app/api/models/news-details.ts
/* tslint:disable */
/* eslint-disable */
export interface NewsDetails {
  content?: null | string;
  detailsTitle?: null | string;
  id?: number;
  image?: null | string;
  mainTitle?: null | string;
  publishDate?: Date;
  urlId?: null | string;
}