# TicketingSystem/ticketing-system/src/app/api/models/news-header.ts
/* tslint:disable */
/* eslint-disable */
export interface NewsHeader {
  id?: number;
  image?: null | string;
  mainTitle?: null | string;
  publishDate?: Date;
  subTitle?: null | string;
  urlId?: null | string;
}