# TicketingSystem/ticketing-system/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TicketingSystemAPI",
    "version": "1.0"
  },
  "servers": [
    {
      "url":"https://localhost:7090"
    }
  ],
  "paths": {
    "/api/Game/games": {
      "get": {
        "tags": [
          "Game"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameHeader"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameHeader"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameHeader"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/{gameId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameHeader"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Game/games/{gameId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game": {
      "post": {
        "tags": [
          "Game"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGame"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGame"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGame"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameHeader"
                }
              }
            }
          }
        }
      }
    },
    "/api/News": {
      "get": {
        "tags": [
          "News"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewsHeader"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewsHeader"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewsHeader"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "News"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNews"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNews"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNews"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/News/{newsUrlId}": {
      "get": {
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "name": "newsUrlId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Player/team": {
      "get": {
        "tags": [
          "Player"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerHeader"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerHeader"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerHeader"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Player/{playerId}": {
      "get": {
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerHeader"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Player": {
      "post": {
        "tags": [
          "Player"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerHeader"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerHeader"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerHeader"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerHeader"
                }
              }
            }
          }
        }
      }
    },
    "/api/Seat/seats": {
      "get": {
        "tags": [
          "Seat"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatHeader"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatHeader"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatHeader"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Seat/{gameId}": {
      "put": {
        "tags": [
          "Seat"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatHeader"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeatHeader"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeatHeader"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SeatHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatHeader"
                }
              }
            }
          }
        }
      }
    },
    "/api/Team": {
      "get": {
        "tags": [
          "Team"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamHeader"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamHeader"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamHeader"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Team"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamHeader"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamHeader"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TeamHeader"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeamHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamHeader"
                }
              }
            }
          }
        }
      }
    },
    "/api/Team/{teamId}": {
      "get": {
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeamHeader"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamHeader"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamHeader"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateGame": {
        "type": "object",
        "properties": {
          "homeTeam": {
            "type": "string",
            "nullable": true
          },
          "awayTeam": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CreateNews": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "urlId": {
            "type": "string",
            "nullable": true
          },
          "mainTitle": {
            "type": "string",
            "nullable": true
          },
          "subTitle": {
            "type": "string",
            "nullable": true
          },
          "detailsTitle": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GameDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "homeTeam": {
            "type": "string",
            "nullable": true
          },
          "awayTeam": {
            "type": "string",
            "nullable": true
          },
          "homeTeamLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "awayTeamLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "place": {
            "type": "string",
            "nullable": true
          },
          "sectors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectorHeader"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GameHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "homeTeam": {
            "type": "string",
            "nullable": true
          },
          "awayTeam": {
            "type": "string",
            "nullable": true
          },
          "homeTeamLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "awayTeamLogoUrl": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "place": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GameSeatHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "seatId": {
            "type": "integer",
            "format": "int32"
          },
          "sector": {
            "type": "string",
            "nullable": true
          },
          "row": {
            "type": "integer",
            "format": "int32"
          },
          "seatNumber": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewsDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "urlId": {
            "type": "string",
            "nullable": true
          },
          "mainTitle": {
            "type": "string",
            "nullable": true
          },
          "detailsTitle": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewsHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "urlId": {
            "type": "string",
            "nullable": true
          },
          "mainTitle": {
            "type": "string",
            "nullable": true
          },
          "subTitle": {
            "type": "string",
            "nullable": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "weigth": {
            "type": "integer",
            "format": "int32"
          },
          "heigth": {
            "type": "integer",
            "format": "int32"
          },
          "jerseyNumber": {
            "type": "integer",
            "format": "int32"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RowHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "rowNumber": {
            "type": "integer",
            "format": "int32"
          },
          "sectorId": {
            "type": "integer",
            "format": "int32"
          },
          "seats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeatHeader"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SeatHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seatNumber": {
            "type": "integer",
            "format": "int32"
          },
          "rowId": {
            "type": "integer",
            "format": "int32"
          },
          "rowNumber": {
            "type": "integer",
            "format": "int32"
          },
          "sectorName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "gameSeats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSeatHeader"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SectorHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "sectorName": {
            "type": "string",
            "nullable": true
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RowHeader"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TeamHeader": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "homeCourt": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}